// src/enums/date-time-format.enum.ts
export enum DateTimeFormat {
    FORMAT_D = "FORMAT_D",
    FORMAT_M = "FORMAT_M",
    FORMAT_M_LONG = "FORMAT_M_LONG",
    FORMAT_M_SHORT = "FORMAT_M_SHORT",
    FORMAT_Y = "FORMAT_Y",
    FORMAT_W = "FORMAT_W",
    FORMAT_DM = "FORMAT_DM",
    FORMAT_MY_LONG = "FORMAT_MY_LONG",
    FORMAT_MY_SHORT = "FORMAT_MY_SHORT",
    FORMAT_WDMY = "FORMAT_WDMY",
    FORMAT_WDM = "FORMAT_WDM",
    FORMAT_DMY = "FORMAT_DMY",
    FORMAT_DATE_TIME = "FORMAT_DATE_TIME",
    FORMAT_TIME = "FORMAT_TIME",
    FORMAT_TIME_WITHOUT_SECONDS = "FORMAT_TIME_WITHOUT_SECONDS"
}